敵をスポーンする処理を分ける
